+++ core/chains/evm/logpoller/log_poller_test.go
@@ -21,6 +21,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"github.com/tendermint/tendermint/libs/rand"
+	"go.uber.org/zap/zapcore"
 
 	"github.com/smartcontractkit/chainlink/v2/core/chains/evm/client"
 	"github.com/smartcontractkit/chainlink/v2/core/chains/evm/logpoller"
@@ -873,4 +874,65 @@ func TestGetReplayFromBlock(t *testing.T) {
 	fromBlock, err = th.LogPoller.GetReplayFromBlock(testutils.Context(t), requested)
 	require.NoError(t, err)
 	assert.Equal(t, requested, fromBlock)
+}
+
+func TestLogPoller_DBErrorHandling(t *testing.T) {
+	t.Parallel()
+	ctx := testutils.Context(t)
+	lggr, observedLogs := logger.TestLoggerObserved(t, zapcore.WarnLevel)
+	chainID1 := testutils.NewRandomEVMChainID()
+	chainID2 := testutils.NewRandomEVMChainID()
+	db := pgtest.NewSqlxDB(t)
+	o := logpoller.NewORM(chainID1, db, lggr, pgtest.NewQConfig(true))
+
+	owner := testutils.MustNewSimTransactor(t)
+	ethDB := rawdb.NewMemoryDatabase()
+	ec := backends.NewSimulatedBackendWithDatabase(ethDB, map[common.Address]core.GenesisAccount{
+		owner.From: {
+			Balance: big.NewInt(0).Mul(big.NewInt(10), big.NewInt(1e18)),
+		},
+	}, 10e6)
+	_, _, emitter, err := log_emitter.DeployLogEmitter(owner, ec)
+	require.NoError(t, err)
+	_, err = emitter.EmitLog1(owner, []*big.Int{big.NewInt(9)})
+	require.NoError(t, err)
+	_, err = emitter.EmitLog1(owner, []*big.Int{big.NewInt(7)})
+	require.NoError(t, err)
+	ec.Commit()
+	ec.Commit()
+	ec.Commit()
+
+	lp := logpoller.NewLogPoller(o, client.NewSimulatedBackendClient(t, ec, chainID2), lggr, 1*time.Hour, 2, 3, 2, 1000)
+
+	err = lp.Replay(ctx, 5) // block number too high
+	require.ErrorContains(t, err, "Invalid replay block number")
+
+	// Force a db error while loading the filters (tx aborted, already rolled back)
+	require.Error(t, utils.JustError(db.Exec(`invalid query`)))
+	go func() {
+		err = lp.Replay(ctx, 2)
+		assert.ErrorContains(t, err, "current transaction is aborted")
+	}()
+
+	time.Sleep(100 * time.Millisecond)
+	lp.Start(ctx)
+	require.Eventually(t, func() bool {
+		return observedLogs.Len() >= 5
+	}, 2*time.Second, 20*time.Millisecond)
+	lp.Close()
+
+	logMsgs := make(map[string]int)
+	for _, obs := range observedLogs.All() {
+		_, ok := logMsgs[obs.Entry.Message]
+		if ok {
+			logMsgs[(obs.Entry.Message)] = 1
+		} else {
+			logMsgs[(obs.Entry.Message)]++
+		}
+	}
+
+	assert.Contains(t, logMsgs, "SQL ERROR")
+	assert.Contains(t, logMsgs, "Failed loading filters in main logpoller loop, retrying later")
+	assert.Contains(t, logMsgs, "Error executing replay, could not get fromBlock")
+	assert.Contains(t, logMsgs, "backup log poller ran before filters loaded, skipping")
 }
